fix(identity): thread enclosing class-template parameter names into member declarations
_walk computed child_template_type_param_names as own-names-if-
FunctionTemplateDecl-else-(), resetting to empty for every other node
kind including a ClassTemplateDecl. An ordinary (non-template) member of
a class template pattern -- template<class T> struct A { void f(T); };
-- is never itself a FunctionTemplateDecl, so it never received the
enclosing class templates own parameter names to canonicalize its
dependent ordinary parameter type against: renaming T to U (the
identical declaration) fingerprinted as a remove+add.

ClassTemplatePartialSpecializationDecl carries the identical shape and
needs the same treatment.

Fixed by making child_template_type_param_names an ACCUMULATING value
(like in_template), not a per-node reset: a ClassTemplateDecl/
ClassTemplatePartialSpecializationDecl nodes own names (via a new
class_template_type_param_names, sharing function_template_type_param_
names exact node-shape walk) are appended to whatever was already
accumulated, and every other node kind inherits the incoming value
unchanged. A member function template nested inside a class template
now sees both levels names.

Split tests/test_entity_id_template_discriminators.py out of
tests/test_entity_id_carrier.py (which had grown past the architecture
gates 1200-line test-file cap) purely to keep both files under it --
no test content changed by the split.
